More tests refactored
