--- Tests/modules/type_related/types_test.py
@@ -1,35 +1,0 @@
-#####################################################################################
-#
-#  Copyright (c) Microsoft Corporation. All rights reserved.
-#
-# This source code is subject to terms and conditions of the Apache License, Version 2.0. A
-# copy of the license can be found in the License.html file at the root of this distribution. If
-# you cannot locate the  Apache License, Version 2.0, please send an email to
-# ironpy@microsoft.com. By using this source code in any fashion, you are agreeing to be bound
-# by the terms of the Apache License, Version 2.0.
-#
-# You must not remove this notice, or any other, from this software.
-#
-#
-#####################################################################################
-
-'''
-This tests what CPythons test_types.py does not hit.
-'''
-
-#--IMPORTS---------------------------------------------------------------------
-from iptest.assert_util import *
-
-import types
-
-#--GLOBALS---------------------------------------------------------------------
-
-#--HELPERS---------------------------------------------------------------------
-
-#--TEST CASES------------------------------------------------------------------
-def test_cp24741():
-    m = types.ModuleType('m')
-    AreEqual(m.__doc__, None)
-
-#--MAIN------------------------------------------------------------------------
-run_test(__name__)
